migrations/versions/99df47d5fb24_initial_schema.py
"""initial schema

Revision ID: 99df47d5fb24
Revises: 
Create Date: 2025-12-08 11:50:54.231467

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '99df47d5fb24'
down_revision: Union[str, Sequence[str], None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
